requirements.txt
Django>=6.0,<7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
cloudinary>=1.36.0
python-dotenv>=1.0.0
